fix: prevent crash when mixer isn't initialized
It seems when the user's system has problems with audio, pygame.init()
won't complain and the problem will only raise an error when a method
of pygame.mixer.music is used.

This was causing the app to crash in such circumstances because the
audio player uses a method from that module on startup to set the
volume.

Now we only set the volume when the user will actually play audio
(from the audio preview widget) and employ a try/except clause that
catches the error if it occurs, informing the user of the problem
and its implications.

The app doesn't crash anymore and users can still use it normally
in such circumstances (except for playing audio, of course) until
they figure out the problem.

Fixes: #108