fix: prevent crash when replacing operation
Operator nodes can have their operation replaced by right-clicking
them and picking another operation in a submenu from the popup that
appears. Internally, this works by actually replacing the operator
node with a new one with the other operation.

When that happens, existing connections are supposed to be established
with the new operator node. There was a bug in this part of the code,
though, which was causing the app to crash if there was any existing
connection.

The bug was caused by a simple missing assignment to an attribute
which was needed in a subsequent step. Such assignment was made in
two different spots in the method.

Adding the attribute assignments fixed the bug.

The corresponding code was also slightly improved for clarity.

Fixes #110