add: catch misleading warning
The pygame-ce library versions 1.5.2 and 1.5.3 has a bug that
causes a needless/misleading warning to be printed when
pygame.display.set_mode() is called with size (0, 0) and the
RESIZABLE flag.

Such bug is explained in issue 3385 of the pygame-ce repo.

Because of that we now suppress such warning when set_mode()
is called under the aforementioned circumstances.