change: replace font files
By the original, unmixed versions. That is, 02 of the main fonts used
had Japanese characters (also Kanji) mixed in, in the hope that it
would be enough for displaying text for both Japanese and Chinese
users.

Now, instead, we'll seek a solution that completely satisfies the
needs of each language.